Fix: 마이챌린지 에러바운더리 분리 (#403)
Related to: #401
