+++ src/components/LogIn/GithubLoginButton/index.tsx
@@ -3,7 +3,7 @@ import Button from "@/components/LogIn/Button/Button";
 
 function GithubLoginButton() {
   const GITHUB_REDIRECT_URI =
-    "https://api.gitget.co.kr/oauth2/authorization/github";
+    "http://localhost:8080/oauth2/authorization/github";
 
   const onClick = () => {
     window.location.href = GITHUB_REDIRECT_URI;
