+++ src/index.css
@@ -4,6 +4,7 @@
 
 html {
   font-size: 62.5%;
+  overflow-y: scroll;
 }
 
 * {
@@ -12,7 +13,6 @@ html {
 
 #root {
   height: 100vh;
-  overflow-y: scroll;
 }
 
 :root {
