+++ src/pages/MyChallenge/MyChallenge.tsx
@@ -1,11 +1,6 @@
-import { Suspense } from "react";
 import { Outlet } from "react-router-dom";
-import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
 import MainHeader from "@/components/Common/MainHeader/MainHeader";
 import MyChallengeTabs from "@/components/Main/MyChallenge/MyChallengeTabs/MyChallengeTabs";
-import { ErrorBoundary } from "react-error-boundary";
-import CommonGetErrorFallback from "@/components/Error/CommonGetErrorFallback/CommonGetErrorFallback";
-import { QueryErrorResetBoundary } from "react-query";
 
 const MyChallenge = () => {
   return (
@@ -16,18 +11,7 @@ const MyChallenge = () => {
       </div>
 
       <div className="px-[2.2rem] pt-[12.5rem] h-full pb-[2rem]">
-        <QueryErrorResetBoundary>
-          {({ reset }) => (
-            <ErrorBoundary
-              onReset={reset}
-              FallbackComponent={CommonGetErrorFallback}
-            >
-              <Suspense fallback={<LoadingBox />}>
-                <Outlet />
-              </Suspense>
-            </ErrorBoundary>
-          )}
-        </QueryErrorResetBoundary>
+        <Outlet />
       </div>
     </>
   );
