+++ src/setupProxy.js
@@ -4,7 +4,7 @@ export default (app) => {
   app.use(
     "/api",
     createProxyMiddleware({
-      target: "https://api.gitget.co.kr",
+      target: "http://localhost:8080",
       changeOrigin: true,
     })
   );
